requirements.txt
requests>=2.31.0
msal>=1.28.0
PyYAML>=6.0
pytest>=8.0
tenacity>=8.2.0
